[WCNCR00276682][Coverity coding defect]
Argument type for printf format specifier

MTK-Commit-Id: 22b690c0586c72837dd8775dd5d4a2f837be97e8

CR-Id: WCNCR00276682
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6914699
CheckSWIP: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_pf_nep_sanity
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>

+++ mgmt/twt_planner.c
@@ -1454,13 +1454,14 @@ void twtPlannerGetTsfDone(
 		DBGLOG(TWT_REQUESTER, WARN,
 		"TWT Info Frame[2] %d Tgt[0x%x 0x%x] Cur[0x%x 0x%x] Input[0x%x 0x%x]\n",
 		ucNextTWTSize,
-		(u8NextTWT & 0x00000000FFFFFFFF),
-		((u8NextTWT & 0xFFFFFFFF00000000) >> 32),
-		(u8CurTsf & 0x00000000FFFFFFFF),
-		((u8CurTsf & 0xFFFFFFFF00000000) >> 32),
-		(prGetTsfCtxt->rNextTWT.u8NextTWT & 0x00000000FFFFFFFF),
-		((prGetTsfCtxt->rNextTWT.u8NextTWT & 0xFFFFFFFF00000000)
-			>> 32));
+		(uint32_t)(u8NextTWT & 0x00000000FFFFFFFF),
+		(uint32_t)((u8NextTWT & 0xFFFFFFFF00000000) >> 32),
+		(uint32_t)(u8CurTsf & 0x00000000FFFFFFFF),
+		(uint32_t)((u8CurTsf & 0xFFFFFFFF00000000) >> 32),
+		(uint32_t)(prGetTsfCtxt->rNextTWT.u8NextTWT &
+				0x00000000FFFFFFFF),
+		(uint32_t)((prGetTsfCtxt->rNextTWT.u8NextTWT &
+				0xFFFFFFFF00000000) >> 32));
 
 		/* Start the process to resume this TWT agreement */
 		twtPlannerSendReqResume(prAdapter,
